+++ CHANGES.md
@@ -8,6 +8,7 @@ New release
 - add meetings filter in Time entries report
 - add bigger than field in Time entries
 - custom twistd launcher - add daily log rotation
+- grouping feature in Time entries to excel
 
 0.8.2 (13-05-2013)
 ===============